new commit
The project file contains a python script (project.py) and database files. This is a simple GUI based project which is very easy to understand and use. Talking about the system, it contains all the basic functions which include adding, viewing, deleting and updating items. In order to add a medicine detail, the user has to provide its name, price, quantity, category, and discount.